# Avoid allocations for sorting in bulk load